+++ src/app/hull-app.js
@@ -3,7 +3,7 @@ import Promise from "bluebird";
 import Server from "./server";
 import Worker from "./worker";
 import { Instrumentation, Cache, Queue, Batcher } from "../infra";
-import { exitHandler, serviceMiddleware, tokenMiddleware, notifMiddleware } from "../utils";
+import { exitHandler, serviceMiddleware, tokenMiddleware, notifMiddleware, segmentsMiddleware, requireHullMiddleware } from "../utils";
 
 
 export default function HullApp({
@@ -24,7 +24,7 @@ export default function HullApp({
   exitHandler(() => {
     return Promise.all([
       Batcher.exit(),
-      this.queue.exit()
+      queue.exit()
     ]);
   });
 
@@ -34,23 +34,29 @@ export default function HullApp({
       this._server.use(tokenMiddleware);
       this._server.use(notifMiddleware());
       this._server.use(Hull.Middleware({ hostSecret, clientConfig }));
+      this._server.use(segmentsMiddleware);
       this._server.use(serviceMiddleware(service));
       return this._server;
     },
     worker: function getWorker() {
       this._worker = new Worker({ Hull, instrumentation, cache, queue });
       this._worker.use(Hull.Middleware({ hostSecret, clientConfig }));
+      this._worker.use(requireHullMiddleware);
+      this._server.use(segmentsMiddleware);
       this._worker.use(serviceMiddleware(service));
       return this._worker;
     },
     start: function startApp({ server = true, worker = false } = {}) {
       if (this._server && server) {
         this._server.use(instrumentation.stopMiddleware());
-        this._server.listen(port);
+        this._server.listen(port, () => {
+          Hull.logger.info("HullApp.server.listen", port);
+        });
       }
 
       if (this._worker && worker) {
         this._worker.process();
+        Hull.logger.info("HullApp.worker.process");
       }
     }
   };