+++ src/helpers/handle-extract.js
@@ -9,19 +9,19 @@ import BatchStream from "batch-stream";
 /**
  * @param {Object} ctx The Context Object
  * @param {Object} body Request Body Object
- * @param {Object} chunkSize
+ * @param {Object} batchSize
  * @param {Function} callback returning a Promise
  * @return {Promise}
  *
  * return handleExtract(req, 100, (users) => Promise.resolve())
  */
-export default function handleExtract(ctx, { body, chunkSize, handler }) {
+export default function handleExtract(ctx, { body, batchSize, handler }) {
   const { logger } = ctx.client;
   const { url, format } = body;
   if (!url) return Promise.reject(new Error("Missing URL"));
   const decoder = format === "csv" ? CSVStream.createStream({ escapeChar: "\"", enclosedChar: "\"" }) : JSONStream.parse();
 
-  const batch = new BatchStream({ size: chunkSize });
+  const batch = new BatchStream({ size: batchSize });
 
   return request({ url })
     .pipe(decoder)