+++ src/infra/queue/enqueue.js
@@ -1,7 +1,12 @@
 export default function enqueue(queueAdapter, ctx, jobName, jobPayload, options = {}) {
+  const { id, secret, organization } = ctx.client.configuration();
   const context = {
     hostname: ctx.hostname,
-    query: ctx.client.configuration()
+    query: {
+      ship: id,
+      secret,
+      organization
+    }
   };
 
   return queueAdapter.create("queueApp", {