+++ src/utils/segments-middleware.js
@@ -1,3 +1,7 @@
+import _ from "lodash";
+
+const fieldPath = "ship.private_settings.synchronized_segments";
+
 /**
  * @param  {Object}   req
  * @param  {Object}   res
@@ -10,7 +14,7 @@ export default function segmentsMiddleware(req, res, next) {
     return next();
   }
 
-  const { message } = req.hull;
+  const { cache, message } = req.hull;
   const bust = (message && (message.Subject === "segment:update" || message.Subject === "segment:delete"));
 
   return (() => {
@@ -20,8 +24,13 @@ export default function segmentsMiddleware(req, res, next) {
     return Promise.resolve();
   })().then(() => {
     return cache.wrap("segments", () => req.hull.client.get("/segments"));
-  ]).then((segments) => {
-    req.hull.segments = segments;
+  }).then((segments) => {
+    req.hull.segments = _.map(segments, (s) => {
+      if (_.has(req.hull, fieldPath)) {
+        s.filtered = _.includes(_.get(req.hull, fieldPath, []), s.id);
+      }
+      return s;
+    });
     return next();
   }, () => next());
 }