+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Adds `helpers` with functions being added to the `req.hull.client` after being initiated by the `Hull.Middleware`
 * Adds `infra` with `Instrumentation`, `Queue`, `Cache` and `Batcher` services
 * Adds `utils` with `handlers`, `middlewares` and `routers` used as a building blocks of the `HullApp` and to be used by the ship/connector directly
+* BREAKING - changed the order of arguments in NotifHandler handlers
 
 
 # 0.10.7